- Refactor by sepearing plugins in different files
